extract enable when associate cti_profile
reason: allow to associate cti_profile without enable it.

username/password/cti_enabled will be deprecated soon but the
cti_profile_id will be not deprecated. So we want to set cti_profile_id
without cti_enabled/username/password

If cti_enabled is set to True and not username/password are provided, an
error is raised